Changed setup.py gensim version.
Instead of allowing versions to be >=, all versions are fixed.
This is to avoid changes in the libraries that can cause fails in the
topics_and_summary library.

For example, gensim 3.7.3 has changed one method from the 3.4.0 that causes
that ngrams_model_funcs fails when they receive 2 words that generate an ngram.